Release 2.0
This is a generic packet to use dkms for patching kernel module sources.
The sources to patch are not included in this package - they are fetched
from kernel.org during dkms runtime.
This package contains the mechanism to enable wake on lan on alx ethernet
devices again. There are two examples for other patches included.